requirements.txt
streamlit==1.39.0
tensorflow==2.10.0
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.2
